monitoring-service/requirements.txt
fastapi
uvicorn
motor
python-socketio
aio-pika
python-dotenv
asyncio
requests
pydantic
prometheus-fastapi-instrumentator
pydantic-settings
paho-mqtt
